requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
dataclasses-json==0.6.7
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.2
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.1
langchain-experimental==0.0.42
langchain-ollama==1.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.4
ollama==0.6.0
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
propcache==0.4.1
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
